Ported the ZooKeeper tests to CMake and Windows.
This adds the ZooKeeper test sources to the CMake build system when Java
is enabled, and makes the tests compatible with the CMake build
structure and Windows.

`Path::basename()` cannot be used for URIs on Windows because the
`os::PATH_SEPARATOR` is a backslash, but URIs unconditionally use
forward slashes.