Added gRPC support in libprocess.
A gRPC client can use `process::grpc::call(...)` to send an asynchronous
gRPC call and get a future for the response. The client needs to set up
two data structures: a `Channel` which represents a connection to a gRPC
server, and a `ClientRuntime` which maintains a `CompletionQueue` that
keeps track of all pending asynchronous gRPC calls, and spawns a thread
waiting for any response from the `CompletionQueue`. All gRPC calls
using the same `ClientRuntime` would share the same thread.

Review: https://reviews.apache.org/r/61097