Printed out the source profile and ID for a disk resource.
`Resource::DiskInfo::Source` is printed out in the following formats:

  TYPE(profile,id):root

If the source has no profile and ID, then "(profile,id)" will be
skipped. If either the profile or the ID is missing, the missing part
will be replaced with an empty string. If the source has no root path,
":root" will be skipped.

Review: https://reviews.apache.org/r/64591