Fixed a corner case for pre-existing volumes created by old RPs.
When an agent failed over and registered as a new one, the RP will be
registered as a new one as well, but it could pick up the checkpointed
states for volumes created by an old RP as pre-existing volumes. We
should use the checkpointed capabilities for these volumes instead of
using the default mount or block capability.

Review: https://reviews.apache.org/r/64621