Rewrote test `ConvertPreExistingVolume` for `CREATE_DISK`.
Due to the changes of the `CREATE_DISK` semantics, this test is
rewritten to convert a preprovisioned volume to a profile volumes, and
then to destroy it to return the space back to the storage pool.

NOTE: The updated test will fail unless r/69361 (which implements the
new `CREATE_DISK` semantics) is also applied.

Review: https://reviews.apache.org/r/69360