Made CSI volume managers no longer maintain a list of services.
Since the service manager is passed in, there is no need for the volume
manager to maintain a separated list of services provided by the service
manager. This patch cleans up this part and make the service manager
exposes the list of services it contains. It also add more validation
in the service manager to make the code more cohesive.

Review: https://reviews.apache.org/r/70441